requirements.txt
# Core Data Science & Plotting
pandas>=1.3.0
numpy>=1.21.0
scikit-learn>=1.0.0
joblib>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
missingno>=0.5.0

# Feature Engineering
matminer
jarvis-tools
mp-api
pyarrow

# Environment & API
python-dotenv
fastapi
uvicorn[standard]